Add optional extra_classes to svg_image functions
Hi! Thanks for this package.

I found I wanted the ability to add an arbitrary class name to the `i`
element that the package generates. I don't necessarily love the key
name `extra_classes` but naming things is hard. I'm happy to change it
to a better name if you have one.

I've tested this locally on the app I'm working on, but since there's no
functioning test suite, it's worth further testing.