Finish implementing pre-compiled svg_image functions
This will precompile the user's images into functions that let them
use images from the default collection, custom collections, and insert
dynamic HTML attributes.

This also cleans up the documentation and changes the tone to imply that
the precompiled functions are the default way to use this library, instead
of the "new" way.

Eventually the "old" way should be deprecated and removed from the
library.